+++ src/main/java/uniandes/edu/co/proyecto/Repositorio/HotelRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import jakarta.transaction.Transactional;
 
-public interface HotelRepository extends JpaRepository<Hotel, Integer> {
+public interface HotelRepository extends JpaRepository<Hotel, Long> {
      @Query(value = "SELECT * FROM Hoteles", nativeQuery = true)
         Collection<Hotel> darHoteles();
