+++ src/main/java/uniandes/edu/co/proyecto/Repositorio/Sala_ExtraRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.data.jpa.repository.Modifying;
 import jakarta.transaction.Transactional;
 
-public interface Sala_ExtraRepository extends JpaRepository<ExtraSala, Integer>{
+public interface Sala_ExtraRepository extends JpaRepository<ExtraSala, Long>{
 @Query(value = "SELECT * FROM Sala_Extra", nativeQuery = true)
     Collection<SalaExtra> darReserva_Servicio();
